+++ lib/generators/css_zero/install/install_generator.rb
@@ -16,6 +16,16 @@ class CssZero::InstallGenerator < Rails::Generators::Base
     append_to_file "config/importmap.rb", %(pin_all_from "app/javascript/helpers", under: "helpers"\n)
   end
 
+  def copy_javascript_initializers
+    copy_file "app/javascript/initializers/index.js"
+  end
+
+  def pin_javascript_initializers
+    return unless install_with_importmap?
+    append_to_file "config/importmap.rb", %(pin_all_from "app/javascript/initializers", under: "initializers"\n)
+    append_to_file "app/javascript/application.js", %(import "initializers"\n)
+  end
+
   private
     def install_with_importmap?
       Rails.root.join("config/importmap.rb").exist?